Reorganize classes and tests
