+++ prime-router/src/test/kotlin/fhirengine/translation/hl7/schema/converter/ConverterSchemaTests.kt
@@ -1,4 +1,4 @@
-package gov.cdc.prime.router.fhirengine.translation.hl7.schema
+package gov.cdc.prime.router.fhirengine.translation.hl7.schema.converter
 
 import assertk.assertAll
 import assertk.assertThat
@@ -9,9 +9,10 @@ import assertk.assertions.isFailure
 import assertk.assertions.isFalse
 import assertk.assertions.isNotEmpty
 import assertk.assertions.isTrue
+import gov.cdc.prime.router.fhirengine.translation.hl7.schema.fhirTransform.FHIRTransformSchemaElement
 import kotlin.test.Test
 
-class ConfigSchemaConverterTests {
+class ConverterSchemaTests {
     @Test
     fun `test validate schema`() {
         var schema = ConverterSchema()
